+++ morphy-cbh/src/main/java/se/yarin/cbhlib/games/GameHeaderStorageBase.java
@@ -56,7 +56,7 @@ abstract class GameHeaderStorageBase {
      * @param movesOffset a game is only affected if its moves offset is greater than this
      * @param insertedBytes the number of bytes to adjust with
      */
-    abstract void adjustMovesOffset(int startGameId, int movesOffset, int insertedBytes);
+    abstract void adjustMovesOffset(int startGameId, long movesOffset, long insertedBytes);
 
     /**
      * Adjusts the annotation offset of all game headers that have their annotation offsets
@@ -65,7 +65,7 @@ abstract class GameHeaderStorageBase {
      * @param annotationOffset a game is only affected if its annotation offset is greater than this
      * @param insertedBytes the number of bytes to adjust with
      */
-    abstract void adjustAnnotationOffset(int startGameId, int annotationOffset, int insertedBytes);
+    abstract void adjustAnnotationOffset(int startGameId, long annotationOffset, long insertedBytes);
 
     /**
      * The number of modifying operations to the storage since it was opened.
